split the part of generating train set to setGenerate
